+++ crates/adapters/sp1/src/lib.rs
@@ -11,9 +11,13 @@ use crypto::{SP1PublicKey, SP1Signature};
 use schemars::JsonSchema;
 use serde::de::DeserializeOwned;
 use serde::{Deserialize, Serialize};
+#[cfg(not(target_os = "zkvm"))]
+use sov_rollup_interface::zk::Proof;
 use sov_rollup_interface::zk::{CodeCommitment, CryptoSpec, ZkVerifier};
 #[cfg(not(target_os = "zkvm"))]
-use sp1_sdk::SP1ProofWithPublicValues;
+use sp1_sdk::blocking::{Prover, ProverClient};
+#[cfg(not(target_os = "zkvm"))]
+use sp1_sdk::{SP1ProofWithPublicValues, SP1PublicValues, SP1VerifyingKey};
 
 #[cfg(feature = "native")]
 use crate::crypto::private_key::SP1PrivateKey;
@@ -93,12 +97,11 @@ impl ZkVerifier for SP1Verifier {
         serialized_proof: &[u8],
         code_commitment: &Self::CodeCommitment,
     ) -> Result<T, Self::Error> {
-        let proof: SP1ProofWithPublicValues = bincode::deserialize(serialized_proof)?;
+        let proof = decode_sp1_proof(serialized_proof)?;
+        let prover = ProverClient::builder().cpu().build();
+        let verifying_key: SP1VerifyingKey = bincode::deserialize(&code_commitment.0)?;
 
-        let prover = sp1_sdk::blocking::ProverClient::from_env();
-        let verifying_key: sp1_sdk::SP1VerifyingKey = bincode::deserialize(&code_commitment.0)?;
-        sp1_sdk::blocking::Prover::verify(&prover, &proof, &verifying_key, None)?;
-
+        Prover::verify(&prover, &proof, &verifying_key, None)?;
         Ok(bincode::deserialize(proof.public_values.as_slice())?)
     }
 }
@@ -135,6 +138,16 @@ impl ZkVerifier for SP1Verifier {
         // which are then read within the verify_sp1_proof method. The `verify_sp1_proof` method takes in the vkey hash, and the public values digest as direct input.
         // In the future, SP1 will support an interface for passing all 3 in directly.
         todo!("Implement this.")
+    }
+}
+
+#[cfg(not(target_os = "zkvm"))]
+fn decode_sp1_proof(serialized_proof: &[u8]) -> Result<SP1ProofWithPublicValues, Error> {
+    match bincode::deserialize::<Proof<SP1ProofWithPublicValues, SP1PublicValues>>(
+        serialized_proof,
+    )? {
+        Proof::Full(proof) => Ok(proof),
+        Proof::PublicData(_) => anyhow::bail!("SP1Verifier supports only full proofs"),
     }
 }